(feat) Live reloading for angular
Many thanks to @joe59 for the tip on live reloading in angular.

It's as simple as adding --watch=true to our dev ng build command. The side effect is that front build runs forever and then rebuilds when it detects a change. A full rebuild can be easily forced with a ```docker-compose --build```

If this passes review, I'll update the minds documentation with the details

https://gitlab.com/minds/front/issues/72